ci/ui: split rancher functions into separate file
