+++ tests/cypress/latest/e2e/unit_tests/elemental_plugin.spec.ts
@@ -12,14 +12,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { TopLevelMenu } from '~/support/toplevelmenu';
+import { Rancher } from '~/support/rancher';
 import '~/support/commands';
 import filterTests from '~/support/filterTests.js';
 
 filterTests(['main', 'upgrade'], () => {
   Cypress.config();
   describe('Install Elemental plugin', () => {
-    const topLevelMenu = new TopLevelMenu();
+    const rancher = new Rancher();
   
     beforeEach(() => {
       cy.login();
@@ -28,15 +28,12 @@ filterTests(['main', 'upgrade'], () => {
   
     it('Add elemental-ui repo', () => {
       if ( Cypress.env('elemental_ui_version') != "stable") {
-        topLevelMenu.openIfClosed();
-        cy.contains('local')
-          .click();
-        cy.addHelmRepo('elemental-ui', 'https://github.com/rancher/elemental-ui.git', 'git');
+        rancher.addRepository('elemental-ui', 'https://github.com/rancher/elemental-ui.git', 'git');
       };
     });
     
     it('Enable extension support', () => {
-      topLevelMenu.openIfClosed();
+      rancher.burgerMenuOpenIfClosed();
       cy.contains('Extensions')
         .click();
       // Make sure we are on the Extensions page
@@ -53,7 +50,7 @@ filterTests(['main', 'upgrade'], () => {
     });
   
     it('Install Elemental plugin', () => {
-      topLevelMenu.openIfClosed();
+      rancher.burgerMenuOpenIfClosed();
       cy.contains('Extensions')
         .click();
       cy.contains('elemental');
