+++ tests/cypress/latest/support/commands.ts
@@ -32,60 +32,7 @@ const hwLabels: hardwareLabels = {
 
 // Generic commands
 
-// Log into Rancher
-Cypress.Commands.add('login', (
-  username = Cypress.env('username'),
-  password = Cypress.env('password'),
-  cacheSession = Cypress.env('cache_session')) => {
-    const login = () => {
-      let loginPath ="/v3-public/localProviders/local*";
-      cy.intercept('POST', loginPath).as('loginReq');
-      
-      cy.visit('/auth/login');
 
-      cy.getBySel('local-login-username')
-        .type(username, {log: false});
-
-      cy.getBySel('local-login-password')
-        .type(password, {log: false});
-
-      cy.getBySel('login-submit')
-        .click();
-      cy.wait('@loginReq');
-      cy.getBySel('banner-title').contains('Welcome to Rancher');
-      } 
-
-    if (cacheSession) {
-      cy.session([username, password], login);
-    } else {
-      login();
-    }
-});
-
-// Search by data-testid selector
-Cypress.Commands.add("getBySel", (selector, ...args) => {
-  return cy.get(`[data-testid=${selector}]`, ...args);
-});
-
-// Search fields by label
-Cypress.Commands.add('byLabel', (label) => {
-  cy.get('.labeled-input')
-    .contains(label)
-    .siblings('input');
-});
-
-// Search button by label
-Cypress.Commands.add('clickButton', (label) => {
-  cy.get('.btn')
-    .contains(label)
-    .click();
-});
-
-// Confirm the delete operation
-Cypress.Commands.add('confirmDelete', () => {
-  cy.getBySel('prompt-remove-confirm-button')
-    .click()
-});
 
 // Make sure we are in the desired menu inside a cluster (local by default)
 // You can access submenu by giving submenu name in the array
@@ -143,31 +90,6 @@ for (const command of ['visit', 'click', 'trigger', 'type', 'clear', 'reload', '
         });
     });
 }; 
-
-// Add Helm repo
-Cypress.Commands.add('addHelmRepo', (repoName, repoUrl, repoType) => {
-  cy.clickClusterMenu(['Apps', 'Repositories'])
-
-  // Make sure we are in the 'Repositories' screen (test failed here before)
-  cy.contains('header', 'Repositories')
-    .should('be.visible');
-  cy.contains('Create')
-    .should('be.visible');
-
-  cy.clickButton('Create');
-  cy.contains('Repository: Create')
-    .should('be.visible');
-  cy.typeValue('Name', repoName);
-  if (repoType === 'git') {
-    cy.contains('Git repository')
-      .click();
-    cy.typeValue('Git Repo URL', repoUrl);
-    cy.typeValue('Git Branch', 'main');
-  } else {
-    cy.typeValue('Index URL', repoUrl);
-  }
-  cy.clickButton('Create');
-});
 
 // Delete all resources from a page
 Cypress.Commands.add('deleteAllResources', () => {
