--- tests/cypress/latest/support/toplevelmenu.ts
@@ -1,44 +1,0 @@
-/*
-Copyright © 2022 - 2023 SUSE LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-export class TopLevelMenu {
-  toggle() {
-    cy.getBySel('top-level-menu', {timeout: 12000})
-      .click();
-  }
-
-  openIfClosed() {
-    cy.get('body').then((body) => {
-      if (body.find('.menu.raised').length === 0) {
-        this.toggle();
-      }
-    });
-  }
-
-  categories() {
-    cy.get('.side-menu .body .category');
-  }
-
-  links() {
-    cy.get('.side-menu .option');
-  }
-
-  clusters(clusterName: string) {
-    cy.get('.clusters .cluster.selector.option').contains(clusterName).click();
-  }
-
-  localization() {
-    cy.get('.locale-chooser');
-  }
-}
